Don't expose userdata in /android/userdata
All folders in /android will be bind-mounted during LXC startup.
But we don't want the real userdata partition to be visible from
that container, so it should only be mounted in /userdata in the rootfs.

Also, simplify the creation and bind-mouting of android-data. /data
is now by default a symlink to /android/data, and the latter will
by bind-mounted in LXC's rootfs during the container's startup.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>